refactor for running in browser
