+++ .gitignore
@@ -2,6 +2,8 @@
 .npm
 node_modules
 .DS_Store
+/browser-buffer.js
+/browser-stream.js
 /index.d.ts
 /index.js
 /index.mjs
