+++ src/index.js
@@ -2,7 +2,6 @@ import { isTemplateResult, TemplateResult } from './template-result.js';
 import { browserStreamTemplateRenderer } from './browser-stream-template-renderer';
 import { DefaultTemplateProcessor } from './default-template-processor.js';
 import { DefaultTemplateResultProcessor } from './default-template-result-processor.js';
-import { polyfillBuffer } from './browser-buffer-polyfill.js';
 import { promiseTemplateRenderer } from './promise-template-renderer.js';
 import { streamTemplateRenderer } from './node-stream-template-renderer.js';
 import { Template } from './template.js';
@@ -34,11 +33,12 @@ export {
 const defaultTemplateProcessor = new DefaultTemplateProcessor();
 const defaultTemplateResultProcessor = new DefaultTemplateResultProcessor();
 const templateCache = new Map();
-const isNode = typeof process !== 'undefined' && typeof process.versions.node !== 'undefined';
-
-if (!isNode) {
-  polyfillBuffer();
-}
+const streamRenderer =
+  typeof process !== 'undefined' &&
+  typeof process.versions !== 'undefined' &&
+  typeof process.versions.node !== 'undefined'
+    ? streamTemplateRenderer
+    : browserStreamTemplateRenderer;
 
 /**
  * Interprets a template literal as an HTML template that can be
@@ -67,9 +67,7 @@ function html(strings, ...values) {
  * @returns { import('stream').Readable | ReadableStream }
  */
 function renderToStream(result, options) {
-  return isNode
-    ? streamTemplateRenderer(result, defaultTemplateResultProcessor, options)
-    : browserStreamTemplateRenderer(result, defaultTemplateResultProcessor, options);
+  return streamRenderer(result, defaultTemplateResultProcessor, options);
 }
 
 /**
