+++ src/promise-template-renderer.js
@@ -1,3 +1,5 @@
+import { Buffer } from 'buffer';
+
 /**
  * A factory for rendering a template result to a string resolving Promise
  *
