--- test/browser.html
@@ -1,21 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <meta http-equiv="X-UA-Compatible" content="ie=edge">
-  <title>Test</title>
-</head>
-<body>
-  <main></main>
-  <script>
-    (async function() {
-      const { html, renderToString } = await import('../browser/index.js');
-      const { ifDefined } = await import('../browser/directives/if-defined.js')
-      const elMain = document.querySelector('main');
-      const attr = 'attr';
-      elMain.innerHTML = await renderToString(html`<div attr=${ifDefined(attr)}>Hello, test!</div>`);
-    })()
-  </script>
-</body>
-</html>
