+++ tsconfig.json
@@ -11,8 +11,8 @@
     "allowSyntheticDefaultImports": true,
     "esModuleInterop": true,
     "isolatedModules": true,
-    "declaration": true
+    "noEmit": true
   },
   "include": ["src"],
-  "exclude": ["node_modules", "src/browser-buffer-polyfill.js", "test"]
+  "exclude": ["node_modules", "test"]
 }
